+++ src/CMakeLists.txt
@@ -117,6 +117,7 @@ target_add_warnings(dlaf.auxiliary)
 add_library(dlaf.eigensolver OBJECT
   eigensolver/backtransformation/mc.cpp
   eigensolver/band_to_tridiag/mc.cpp
+  eigensolver/bt_band_to_tri/mc.cpp
   eigensolver/gen_to_std/mc.cpp
   $<$<BOOL:${DLAF_WITH_CUDA}>:eigensolver/gen_to_std/gpu.cpp>
   eigensolver/reduction_to_band/mc.cpp)
