+++ src/eigensolver/backtransformation/mc.cpp
@@ -1,19 +1,10 @@
 #include "dlaf/eigensolver/backtransformation/mc.h"
-#include "dlaf/eigensolver/backtransformation/impl-t2b.h"
 
-namespace dlaf {
-namespace eigensolver {
-namespace internal {
+namespace dlaf::eigensolver::internal {
 
 DLAF_EIGENSOLVER_BACKTRANSFORMATION_MC_ETI(, float)
 DLAF_EIGENSOLVER_BACKTRANSFORMATION_MC_ETI(, double)
 DLAF_EIGENSOLVER_BACKTRANSFORMATION_MC_ETI(, std::complex<float>)
 DLAF_EIGENSOLVER_BACKTRANSFORMATION_MC_ETI(, std::complex<double>)
 
-DLAF_EIGENSOLVER_BACKTRANSFORMATION_T2B_MC_ETI(, float)
-DLAF_EIGENSOLVER_BACKTRANSFORMATION_T2B_MC_ETI(, double)
-DLAF_EIGENSOLVER_BACKTRANSFORMATION_T2B_MC_ETI(, std::complex<float>)
-DLAF_EIGENSOLVER_BACKTRANSFORMATION_T2B_MC_ETI(, std::complex<double>)
-}
-}
 }
